+++ pkg/engram/experiments_test.go
@@ -1,0 +1,9 @@
+package engram
+
+import "testing"
+
+func TestExperimentRegistryIsValid(t *testing.T) {
+	if err := ValidateExperimentRegistry(); err != nil {
+		t.Fatal(err)
+	}
+}